fix(api): rename users package to identity
