+++ services/api/registry_credentials_test.go
@@ -3,11 +3,11 @@ package main
 import (
 	"testing"
 
-	"mcp-sentinel-api/users"
+	"mcp-sentinel-api/identity"
 )
 
 func TestRegistryCredentialUsernameUsesNamespace(t *testing.T) {
-	got := users.RegistryCredentialUsername(principal{
+	got := identity.RegistryCredentialUsername(principal{
 		Namespace: "mcp-team-acme",
 		Subject:   "user-1",
 		Email:     "user@example.com",
@@ -18,7 +18,7 @@ func TestRegistryCredentialUsernameUsesNamespace(t *testing.T) {
 }
 
 func TestRegistryCredentialUsernameFallsBackToSubject(t *testing.T) {
-	got := users.RegistryCredentialUsername(principal{
+	got := identity.RegistryCredentialUsername(principal{
 		Subject: "user-1",
 		Email:   "user@example.com",
 	})
@@ -28,7 +28,7 @@ func TestRegistryCredentialUsernameFallsBackToSubject(t *testing.T) {
 }
 
 func TestRegistryCredentialUsernameFallsBackToEmail(t *testing.T) {
-	got := users.RegistryCredentialUsername(principal{
+	got := identity.RegistryCredentialUsername(principal{
 		Email: "user@example.com",
 	})
 	if got != "user@example.com" {
